+++ src/components/ErrorPage.jsx
@@ -4,6 +4,7 @@ import styles from "../styles/ErrorPage.module.css";
 const ErrorPage = () => {
   const { error, status } = useRouteError();
   console.log(error);
+  console.log(status);
   return (
     <section id={styles["error-section"]} role="region">
       <p>{error.message}</p>